Extracted `Config` class out of `Context`
