+++ zappend/config/__init__.py
@@ -5,17 +5,18 @@
 from .attrs import eval_dyn_config_attrs
 from .attrs import has_dyn_config_attrs
 from .attrs import get_dyn_config_attrs_env
-from .config import ConfigItem
-from .config import ConfigLike
-from .config import ConfigList
-from .config import exclude_from_config
-from .config import merge_configs
-from .config import normalize_config
-from .config import validate_config
+from .config import Config
 from .defaults import DEFAULT_APPEND_DIM
 from .defaults import DEFAULT_APPEND_STEP
 from .defaults import DEFAULT_ATTRS_UPDATE_MODE
 from .defaults import DEFAULT_SLICE_POLLING_INTERVAL
 from .defaults import DEFAULT_SLICE_POLLING_TIMEOUT
 from .defaults import DEFAULT_ZARR_VERSION
+from .normalize import ConfigItem
+from .normalize import ConfigLike
+from .normalize import ConfigList
+from .normalize import exclude_from_config
+from .normalize import merge_configs
+from .normalize import normalize_config
 from .schema import get_config_schema
+from .validate import validate_config
